feat: [pipe-20716]: fix routing & move files to repo
