+++ packages/ui/src/views/index.ts
@@ -12,7 +12,7 @@ export * from './auth'
 export * from './empty-page'
 
 //pull-request-list
-export * from './pull-request'
+export * from './repo/pull-request'
 
 //Webhook
-export * from './webhooks'
+export * from './repo/webhooks'
