+++ packages/ui/src/views/layouts/PullRequestCompareLayout.tsx
@@ -19,12 +19,12 @@ import {
 } from '@components/index'
 import { DiffModeEnum } from '@git-diff-view/react'
 import { zodResolver } from '@hookform/resolvers/zod'
-import PullRequestCompareButton from '@views/pull-request/compare/components/pull-request-compare-button'
-import PullRequestCompareForm from '@views/pull-request/compare/components/pull-request-compare-form'
-import TabTriggerItem from '@views/pull-request/compare/components/pull-request-compare-tab-trigger-item'
+import PullRequestCompareButton from '@views/repo/pull-request/compare/components/pull-request-compare-button'
+import PullRequestCompareForm from '@views/repo/pull-request/compare/components/pull-request-compare-form'
+import TabTriggerItem from '@views/repo/pull-request/compare/components/pull-request-compare-tab-trigger-item'
 // import { useDiffConfig } from '@views/pull-request/hooks/useDiffConfig'
-import { TypesCommit } from '@views/pull-request/types'
-import { parseStartingLineIfOne } from '@views/pull-request/utils'
+import { TypesCommit } from '@views/repo/pull-request/types'
+import { parseStartingLineIfOne } from '@views/repo/pull-request/utils'
 import { z } from 'zod'
 
 import { SandboxLayout } from '..'
