+++ packages/ui/src/views/repo/webhooks/index.ts
@@ -1,0 +1,3 @@
+// webhook list
+export * from '@views/repo/webhooks/webhook-list/repo-webhook-list-page'
+export * from '@views/repo/webhooks/webhook-list/types'
