--- packages/ui/src/views/webhooks/index.ts
@@ -1,3 +1,0 @@
-// webhook list
-export * from '@views/webhooks/webhook-list/repo-webhook-list-page'
-export * from '@views/webhooks/webhook-list/types'
